back/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.7
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==6.0.0
Django==4.2.13
django-allauth==0.61.1
django-extensions==3.2.3
djangorestframework==3.15.1
exceptiongroup==1.2.1
executing==2.0.1
idna==3.7
ipython==8.18.1
jedi==0.19.1
matplotlib-inline==0.1.7
oauthlib==3.2.2
parso==0.8.4
prompt-toolkit==3.0.43
pure-eval==0.2.2
pycparser==2.22
Pygments==2.18.0
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
